test: move the files into correct package
The test files for Gradle were not in the correct package, instead sitting at just LintRules, move t
hem into lintrules.gradle
